refactor(fragmenter): publicize, move method back
