+++ src/frontend/src/stream_fragmenter/rewrite/mod.rs
@@ -13,4 +13,4 @@
 // limitations under the License.
 
 mod delta_join;
-pub(super) use delta_join::build_delta_join_without_arrange;
+pub(crate) use delta_join::build_delta_join_without_arrange;
